Revert ESP32 SDK uplift (#496)
A critical issue was found during release testing - Wifi Scan is
misbehaving

Reverting on 1.4 to unblock release.

Use with https://github.com/WildernessLabs/Meadow-ESP32/tree/meadow-1.4
- this contains the revert on the ESP32 firmware side.

+++ esp32/components/F7Services/Messaging/Encoders.cpp
@@ -454,7 +454,7 @@ void Encoders::EncodeMessage(Message *message, uint8_t *buffer, bool headerOnly)
 {
     uint32_t messageSize = EncodedPacketSize(message, headerOnly);
     uint32_t bufferSize = Encoders::CalculateSpiBufferSize(messageSize);
-    TRACE_MESSAGE_SPECIFY_COMPONENT(SpiTransportProvider::COMPONENT_NAME, "Encoding %u bytes as a %u byte packet", (unsigned int) messageSize, (unsigned int) bufferSize);
+    TRACE_MESSAGE_SPECIFY_COMPONENT(SpiTransportProvider::COMPONENT_NAME, "Encoding %d bytes as a %d byte packet", messageSize, bufferSize);
 
     if (buffer)
     {
